removed unnecessary functions from big class
